# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Chat application API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/conversations/{id}/messages:
    get:
      operationId: adminListMessages
      tags: [admin]
      summary: Full message history of any conversation (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
...
  /admin/conversations/{id}/reply:
    post:
      operationId: adminReplyConversation
      tags: [admin]
      summary: Admin sends a message into any conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Created message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
...
components:
  schemas:
...
    MessageInput:
      type: object
      properties:
        content:
          type: string
        attachmentUrl:
          type: string
        attachmentType:
          type: string
        attachmentName:
          type: string
        encrypted:
          type: boolean
        envelopes:
          type: array
          items:
            $ref: "#/components/schemas/MessageEnvelopeInput"
    MessageEnvelopeInput:
      type: object
      required: [userId, type, body]
      properties:
        userId:
          type: integer
        type:
          type: integer
        body:
          type: string
    Message:
      type: object
      required: [id, conversationId, senderId, isAi, status, createdAt]
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        senderId:
          type: ["integer", "null"]
        senderName:
          type: ["string", "null"]
        isAi:
          type: boolean
        content:
          type: ["string", "null"]
        contentEn:
          type: ["string", "null"]
        encrypted:
          type: boolean
        envelope:
          oneOf:
            - $ref: "#/components/schemas/MessageEnvelopeOut"
            - type: "null"
        attachmentUrl:
          type: ["string", "null"]
        attachmentType:
          type: ["string", "null"]
        attachmentName:
          type: ["string", "null"]
        status:
          type: string
          enum: [sent, delivered, read]
        createdAt:
          type: string
    MessageEnvelopeOut:
      type: object
      required: [type, body]
      properties:
        type:
          type: integer
        body:
          type: string
...